backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
docker==6.1.3
aiofiles==23.2.1
markdown==3.5
python-dotenv==1.0.0
